feat: 스크롤막는 기능 hooks로 분리 #63
