+++ src/components/ui/Modal/index.tsx
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom';
 
 import { useDrag } from '@/hooks/useDrag';
 import * as S from './Modal.styles';
+import { useScrollLock } from '@/hooks/useScrollLock';
 
 type ModalProps = {
   isOpen: boolean;
@@ -17,35 +18,7 @@ const Modal = ({ isOpen, onClose, height, children }: ModalProps) => {
   });
   const modalRoot = document.getElementById('modal-overlay');
 
-  // 모달이 열릴 때 스크롤 방지
-  const preventScroll = () => {
-    const currentScrollY = window.scrollY;
-    const hasScrollBar =
-      window.innerWidth > document.documentElement.clientWidth;
-
-    document.documentElement.style.setProperty(
-      '--scroll-position',
-      `-${currentScrollY}px`,
-    );
-    document.body.classList.add('scroll-locked');
-    if (hasScrollBar) document.body.classList.add('has-scrollbar');
-
-    return currentScrollY;
-  };
-
-  // 모달이 닫힐 때 스크롤 허용
-  const allowScroll = (scrollY: number) => {
-    document.body.classList.remove('scroll-locked', 'has-scrollbar');
-    document.documentElement.style.removeProperty('--scroll-position');
-    window.scrollTo(0, scrollY);
-  };
-
-  useEffect(() => {
-    if (isOpen) {
-      const scrollY = preventScroll();
-      return () => allowScroll(scrollY);
-    }
-  }, [isOpen]);
+  useScrollLock({ isOpen });
 
   useEffect(() => {
     const handleMouseLeave = handlers.onMouseUp;
